(GH-5) Add --sitedir option to commands
Previously the site configuration directory must always be in the users home
directory, however there may be cases where users may want to use a non-default
location, specifically for the unit/integration testing of this project.

This commit adds a `sitedir` option to all commands and sets the default to the
previous value of '~/.localhost/sites'.