(GH-8) Fix npm run script command on Windows
Previously when using the `localhost run` command, on Windows, and the script to
run is like `npm run server`, an error would be thrown indicating npm does not
exist. But running the `npm run server` command manually would work.  This is
due to the spawn command not using a shell to execute the command and that npm
is actually a batch file on Windows not an executable file.  This commit
adds an option to `spawn` to use a shell when executing commands on Windows
platforms.  There may be cases where this applicable to non-Windows platforms
but that has yet to come about.