(GH-7) Add Test fixtures
This commit adds an example project which can be consumed as part of an
integration test suite.  Two scripts are added: index.js which outputs a variety
of message levels to test STDOUT and STDERR redirection and tcpecho.js which
can be used in integration tests to verify TCP tunnelling.